app/requirements.txt
alembic==1.6.5
appdirs==1.4.4
bidict==0.21.2
black==21.7b0
click==8.0.1
distlib==0.3.2
eventlet==0.17.4
filelock==3.0.12
Flask==0.12.2
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==18.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.26
pathspec==0.9.0
protobuf==3.17.3
psycopg2-binary==2.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
alembic==1.6.5
click==8.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-editor==1.0.4
six==1.16.0
Werkzeug==2.0.1

python-engineio==4.2.1
python-socketio==5.4.0
regex==2021.7.6
six==1.16.0
SQLAlchemy==1.4.22
tomli==1.1.0
virtualenv==20.4.7
Werkzeug==2.0.1
